recovery_app/requirements.txt
streamlit==1.47.1
pandas
requests
beautifulsoup4
openpyxl
pyarrow
pillow
watchdog
fastapi
uvicorn

